refactor(workshop): normalize feature slices
