+++ .todo/epics/epic-workshop-architecture-refactor-2026-08-03/epic-workshop-architecture-refactor.md
@@ -2,7 +2,7 @@
 
 **Created:** 2026-08-03
 
-**Status:** In progress — Phase 0
+**Status:** Phase 1 implementation complete — awaiting review
 
 **Priority:** Critical — blocks all new Workshop feature development
