+++ packages/core/src/__tests__/presentation/webview/hooks/domain/useWorkshop.test.ts
@@ -37,8 +37,7 @@ import type {
   WorkshopSessionsDataMessage,
   WorkshopSessionStateMessage,
   WorkshopTurn,
-  WorkshopTurnMessage,
-  WorkshopWidgetActionResultMessage
+  WorkshopTurnMessage
 } from '@messages';
 import { createMockVSCode } from '@/__tests__/mocks/vscode';
 
@@ -425,27 +424,6 @@ describe('useWorkshop', () => {
 
     act(() => result.current.consumeSessionActionResult());
     expect(result.current.sessionActionResult).toBeUndefined();
-  });
-
-  it('owns Gesture commit acknowledgements until the modal consumes them', () => {
-    const { result } = renderHook(() => useWorkshop());
-    const response: WorkshopWidgetActionResultMessage = {
-      type: MessageType.WORKSHOP_WIDGET_ACTION_RESULT,
-      source: 'extension.workshop.widget',
-      payload: {
-        action: 'commit',
-        widgetId: 'gesture-playground',
-        ok: false,
-        message: 'The session is still saving.'
-      },
-      timestamp: 0
-    };
-
-    act(() => result.current.handleWidgetActionResult(response));
-    expect(result.current.widgetActionResult).toEqual(response.payload);
-
-    act(() => result.current.consumeWidgetActionResult());
-    expect(result.current.widgetActionResult).toBeNull();
   });
 
   it('submits one complete conversation behavior object without optimistic state', () => {
@@ -1134,78 +1112,4 @@ describe('useWorkshop', () => {
     });
   });
 
-  describe('Gesture Playground generation progress', () => {
-    it('tracks token-keyed progress and clears it when that result settles', () => {
-      const { result } = renderHook(() => useWorkshop());
-
-      act(() => result.current.handleWidgetGenerationProgress({
-        type: MessageType.WORKSHOP_WIDGET_GENERATION_PROGRESS,
-        source: 'extension.workshop',
-        payload: {
-          widgetId: 'gesture-playground',
-          token: 'gesture-1',
-          phase: 'streaming',
-          stage: 'dictionary',
-          outputCharacters: 4_000,
-          estimatedOutputTokens: 1_000,
-          outputTokenLimit: 50_000
-        },
-        timestamp: 1
-      }));
-
-      expect(result.current.widgetGenerationProgress).toMatchObject({
-        token: 'gesture-1',
-        stage: 'dictionary',
-        estimatedOutputTokens: 1_000
-      });
-
-      act(() => result.current.handleWidgetMenuResult({
-        type: MessageType.WORKSHOP_WIDGET_MENU_RESULT,
-        source: 'extension.workshop',
-        payload: {
-          widgetId: 'gesture-playground',
-          token: 'gesture-1',
-          mode: 'full',
-          ok: false,
-          error: 'Stopped.'
-        },
-        timestamp: 2
-      }));
-
-      expect(result.current.widgetGenerationProgress).toBeNull();
-    });
-
-    it('does not let a stale menu result clear newer progress', () => {
-      const { result } = renderHook(() => useWorkshop());
-
-      act(() => result.current.handleWidgetGenerationProgress({
-        type: MessageType.WORKSHOP_WIDGET_GENERATION_PROGRESS,
-        source: 'extension.workshop',
-        payload: {
-          widgetId: 'gesture-playground',
-          token: 'gesture-new',
-          phase: 'streaming',
-          stage: 'menu',
-          outputCharacters: 8_000,
-          estimatedOutputTokens: 2_000,
-          outputTokenLimit: 50_000
-        },
-        timestamp: 1
-      }));
-      act(() => result.current.handleWidgetMenuResult({
-        type: MessageType.WORKSHOP_WIDGET_MENU_RESULT,
-        source: 'extension.workshop',
-        payload: {
-          widgetId: 'gesture-playground',
-          token: 'gesture-old',
-          mode: 'full',
-          ok: false,
-          error: 'Stale.'
-        },
-        timestamp: 2
-      }));
-
-      expect(result.current.widgetGenerationProgress?.token).toBe('gesture-new');
-    });
-  });
 });
