+++ packages/core/src/application/services/workshop/WorkshopPromptBuilder.ts
@@ -11,7 +11,6 @@ import {
   WorkshopConversationBehaviorTransition,
   WorkshopExcerpt,
   WorkshopExcerptSource,
-  WorkshopGestureDraft,
   WorkshopPersonaId,
   WorkshopTodoItem,
   WorkshopToolId
@@ -638,32 +637,5 @@ export function buildWorkshopHostMessage(
     options.activationFrame ? '' : undefined,
     'WRITER MESSAGE:',
     safeWriterMessage
-  ].filter((line): line is string => line !== undefined).join('\n');
-}
-
-/**
- * Build the compact room directive for one committed Gesture Playground
- * draft. This is application-level prompt assembly; provider infrastructure
- * is deliberately unaware of the room-facing artifact format.
- */
-export function buildGestureDirective(input: Pick<
-  WorkshopGestureDraft,
-  | 'targetPhrase'
-  | 'selections'
-  | 'note'
-  | 'dictionaryMarkdown'
-  | 'includeDictionaryInCommit'
->): string {
-  return [
-    `Gesture directions I want for "${input.targetPhrase.trim()}":`,
-    ...input.selections.map((selection) => `· ${selection}`),
-    input.note.trim().length > 0 ? `note: ${input.note.trim()}` : undefined,
-    ...(input.includeDictionaryInCommit
-      ? [
-          '',
-          'Full Gesture Dictionary shared by the writer as reference:',
-          input.dictionaryMarkdown.trim()
-        ]
-      : [])
   ].filter((line): line is string => line !== undefined).join('\n');
 }
