+++ packages/core/src/application/services/workshop/WorkshopSessionCheckpointNormalization.ts
@@ -11,7 +11,7 @@ import type {
 } from '@/application/services/workshop/WorkshopSessionStateV1';
 import {
   normalizeGesturePlaygroundDraftForHydration
-} from '@/application/services/workshop/widgets/GesturePlaygroundConfigCodec';
+} from '@/application/services/workshop/widgets/gesturePlayground/GesturePlaygroundConfigCodec';
 
 export type WorkshopSessionCheckpointNormalization =
   | 'discarded-legacy-scope-transition'
