+++ packages/core/src/application/services/workshop/widgets/lexicalGravity/LexicalGravityConfigCodec.ts
