+++ packages/core/src/application/services/workshop/widgets/lexicalGravity/LexicalGravityDirective.ts
