+++ packages/core/src/application/services/workshop/widgets/lexicalGravity/LexicalGravityLenses.ts
