+++ packages/core/src/index.ts
@@ -99,7 +99,9 @@ export { LexicalGravityModelService } from '@services/widgets/LexicalGravityMode
 export { LexicalGravityLensRepository } from '@/infrastructure/storage/LexicalGravityLensRepository';
 export { WorkshopWriterProfileService } from '@/application/services/workshop/WorkshopWriterProfileService';
 export { GesturePlaygroundService } from '@services/widgets/GesturePlaygroundService';
-export { buildGestureDirective } from '@/application/services/workshop/WorkshopPromptBuilder';
+export {
+  buildGestureDirective
+} from '@/application/services/workshop/widgets/gesturePlayground/GesturePlaygroundDirective';
 export {
   WORKSHOP_WIDGET_RECOMMENDATION_INSTRUCTION,
   inspectWorkshopWidgetRecommendation,
