+++ packages/core/src/presentation/webview/hooks/domain/useWorkshop.ts
@@ -66,18 +66,7 @@ import {
   WorkshopTurnMessage,
   WorkshopWriterProfile,
   WorkshopWebResearchSettings,
-  WorkshopWidgetConfigSummary,
   WorkshopStandingDirectiveSummary,
-  WorkshopWidgetConfigDataMessage,
-  WorkshopWidgetConfigSnapshot,
-  WorkshopWidgetGeneratePayload,
-  WorkshopWidgetGenerationProgressMessage,
-  WorkshopWidgetGenerationProgressPayload,
-  WorkshopWidgetMenuResultMessage,
-  WorkshopWidgetMenuResultPayload,
-  WorkshopCommitWidgetPayload,
-  WorkshopWidgetActionResultMessage,
-  WorkshopWidgetActionResultPayload,
   coerceWorkshopConversationBehavior,
   coerceWorkshopWriterProfile,
   coerceWorkshopWebResearchSettings
@@ -134,23 +123,8 @@ export interface WorkshopState {
   contextAttachments: WorkshopContextAttachmentSnapshot[];
   /** Staged one-shot attachments for the writer's next message (Phase 6B). */
   pendingMessageAttachments: WorkshopMessageAttachmentSnapshot[];
-  /**
-   * Persisted widget authoring configs by stable `wc-N` id (ADR 2026-07-22) —
-   * host truth mirrored from the session snapshot; chips re-hydrate from it.
-   */
-  widgetConfigs: WorkshopWidgetConfigSummary[];
   /** Active host-owned prose directives for the standing rail. */
   standingDirectives: WorkshopStandingDirectiveSummary[];
-  /** Full config fetched only when a committed widget chip is opened. */
-  widgetConfigData: WorkshopWidgetConfigSnapshot | null;
-  widgetConfigResponseId: string | null;
-  widgetConfigError: string | null;
-  /** Latest generate result; the modal drops results whose token is stale. */
-  widgetMenuResult: WorkshopWidgetMenuResultPayload | null;
-  /** Token-keyed live progress for the private pre-commit model call. */
-  widgetGenerationProgress: WorkshopWidgetGenerationProgressPayload | null;
-  /** Host acknowledgement for Gesture Playground's commit handoff. */
-  widgetActionResult: WorkshopWidgetActionResultPayload | null;
   contextPending: boolean;
   /** Configured resource catalog for the Context Selector; null until requested. */
   contextCatalog: WorkshopContextCatalogEntry[] | null;
@@ -277,16 +251,6 @@ export interface WorkshopActions {
   deleteSession: (sessionId: string) => void;
   consumeSessionActionResult: () => void;
   clearError: () => void;
-  generateWidgetMenu: (payload: WorkshopWidgetGeneratePayload) => void;
-  cancelWidgetGenerate: (requestId: string) => void;
-  requestWidgetConfig: (configId: string) => void;
-  clearWidgetConfigData: () => void;
-  commitWidget: (payload: WorkshopCommitWidgetPayload) => void;
-  handleWidgetMenuResult: (message: WorkshopWidgetMenuResultMessage) => void;
-  handleWidgetConfigData: (message: WorkshopWidgetConfigDataMessage) => void;
-  handleWidgetGenerationProgress: (message: WorkshopWidgetGenerationProgressMessage) => void;
-  handleWidgetActionResult: (message: WorkshopWidgetActionResultMessage) => void;
-  consumeWidgetActionResult: () => void;
   handleSessionState: (message: WorkshopSessionStateMessage) => void;
   handleSessionsData: (message: WorkshopSessionsDataMessage) => void;
   handleSessionActionResult: (message: WorkshopSessionActionResultMessage) => void;
@@ -333,19 +297,8 @@ export const useWorkshop = (): UseWorkshopReturn => {
     React.useState<WorkshopAttachmentContentState | null>(null);
   const [contextAttachments, setContextAttachments] = React.useState<WorkshopContextAttachmentSnapshot[]>([]);
   const [pendingMessageAttachments, setPendingMessageAttachments] = React.useState<WorkshopMessageAttachmentSnapshot[]>([]);
-  const [widgetConfigs, setWidgetConfigs] = React.useState<WorkshopWidgetConfigSummary[]>([]);
   const [standingDirectives, setStandingDirectives] =
     React.useState<WorkshopStandingDirectiveSummary[]>([]);
-  const [widgetConfigData, setWidgetConfigData] =
-    React.useState<WorkshopWidgetConfigSnapshot | null>(null);
-  const [widgetConfigResponseId, setWidgetConfigResponseId] = React.useState<string | null>(null);
-  const [widgetConfigError, setWidgetConfigError] = React.useState<string | null>(null);
-  const [widgetMenuResult, setWidgetMenuResult] =
-    React.useState<WorkshopWidgetMenuResultPayload | null>(null);
-  const [widgetGenerationProgress, setWidgetGenerationProgress] =
-    React.useState<WorkshopWidgetGenerationProgressPayload | null>(null);
-  const [widgetActionResult, setWidgetActionResult] =
-    React.useState<WorkshopWidgetActionResultPayload | null>(null);
   const [contextPending, setContextPending] = React.useState(false);
   const [contextCatalog, setContextCatalog] = React.useState<WorkshopContextCatalogEntry[] | null>(null);
   const [wizardRun, setWizardRun] = React.useState<string | null>(null);
@@ -524,72 +477,6 @@ export const useWorkshop = (): UseWorkshopReturn => {
   const runContextWizard = React.useCallback(() => {
     post(MessageType.WORKSHOP_RUN_CONTEXT_WIZARD, {});
   }, [post]);
-
-  // ── Conversation Widgets (ADR 2026-07-22) ──────────────────────────────────
-
-  const generateWidgetMenu = React.useCallback((payload: WorkshopWidgetGeneratePayload) => {
-    setWidgetMenuResult(null);
-    setWidgetGenerationProgress(null);
-    post(MessageType.WORKSHOP_WIDGET_GENERATE, payload);
-  }, [post]);
-
-  const cancelWidgetGenerate = React.useCallback((requestId: string) => {
-    vscode.postMessage(
-      createCancelRequestMessage('workshop-widget', requestId, 'webview.workshop.widget')
-    );
-  }, [vscode]);
-
-  const requestWidgetConfig = React.useCallback((configId: string) => {
-    setWidgetConfigData(null);
-    setWidgetConfigResponseId(null);
-    setWidgetConfigError(null);
-    post(MessageType.WORKSHOP_REQUEST_WIDGET_CONFIG, { configId });
-  }, [post]);
-
-  const clearWidgetConfigData = React.useCallback(() => {
-    setWidgetConfigData(null);
-    setWidgetConfigResponseId(null);
-    setWidgetConfigError(null);
-  }, []);
-
-  const commitWidget = React.useCallback((payload: WorkshopCommitWidgetPayload) => {
-    setWidgetActionResult(null);
-    post(MessageType.WORKSHOP_COMMIT_WIDGET, payload);
-  }, [post]);
-
-  const handleWidgetMenuResult = React.useCallback(
-    (message: WorkshopWidgetMenuResultMessage) => {
-      setWidgetMenuResult(message.payload);
-      setWidgetGenerationProgress((current) =>
-        current?.token === message.payload.token ? null : current
-      );
-    },
-    []
-  );
-
-  const handleWidgetGenerationProgress = React.useCallback(
-    (message: WorkshopWidgetGenerationProgressMessage) => {
-      setWidgetGenerationProgress(message.payload);
-    },
-    []
-  );
-
-  const handleWidgetActionResult = React.useCallback(
-    (message: WorkshopWidgetActionResultMessage) => {
-      if (
-        message.payload.action === 'commit'
-        && message.payload.widgetId === 'gesture-playground'
-      ) {
-        setWidgetActionResult(message.payload);
-      }
-    },
-    []
-  );
-
-  const consumeWidgetActionResult = React.useCallback(
-    () => setWidgetActionResult(null),
-    []
-  );
 
   const cancelContextWizard = React.useCallback(() => {
     if (wizardRun) {
@@ -817,7 +704,6 @@ export const useWorkshop = (): UseWorkshopReturn => {
       setRoomHasMemory(session.roomHasMemory);
       setContextAttachments(session.contextAttachments ?? []);
       setPendingMessageAttachments(session.pendingMessageAttachments ?? []);
-      setWidgetConfigs(session.widgetConfigs ?? []);
       setStandingDirectives(session.standingDirectives ?? []);
       setContextPending(session.pendingHostUpdate?.context ?? false);
       setTotalTurns(session.totalTurns);
@@ -857,15 +743,6 @@ export const useWorkshop = (): UseWorkshopReturn => {
       }
     },
     [setLiveRun, streaming]
-  );
-
-  const handleWidgetConfigData = React.useCallback(
-    (message: WorkshopWidgetConfigDataMessage) => {
-      setWidgetConfigResponseId(message.payload.configId);
-      setWidgetConfigData(message.payload.config ?? null);
-      setWidgetConfigError(message.payload.error ?? null);
-    },
-    []
   );
 
   const handleSessionsData = React.useCallback((message: WorkshopSessionsDataMessage) => {
@@ -1047,14 +924,7 @@ export const useWorkshop = (): UseWorkshopReturn => {
     attachmentContent,
     contextAttachments,
     pendingMessageAttachments,
-    widgetConfigs,
     standingDirectives,
-    widgetConfigData,
-    widgetConfigResponseId,
-    widgetConfigError,
-    widgetMenuResult,
-    widgetGenerationProgress,
-    widgetActionResult,
     contextPending,
     contextCatalog,
     contextSearch,
@@ -1119,16 +989,6 @@ export const useWorkshop = (): UseWorkshopReturn => {
     clearContextSearch,
     addContextResources,
     attachMessageResources,
-    generateWidgetMenu,
-    cancelWidgetGenerate,
-    requestWidgetConfig,
-    clearWidgetConfigData,
-    commitWidget,
-    handleWidgetMenuResult,
-    handleWidgetConfigData,
-    handleWidgetGenerationProgress,
-    handleWidgetActionResult,
-    consumeWidgetActionResult,
     attachMessageFile,
     removeMessageAttachment,
     setExcerptResource,
